Practica_4/services/telemetry-generator/requirements.txt
# Python dependencies для Telemetry Generator
# Минимальный набор зависимостей для production

psycopg2-binary==2.9.9  # PostgreSQL adapter

